taskcontroller and task
adding of user id to product and restricting other users from edit and deleting other users products

+++ app/Http/Controllers/TasksController.php
@@ -44,6 +44,11 @@ class TasksController extends Controller
 
     public function edit(Task $task)
     {
+
+        if ($task->user_id != auth()->id()){
+            abort(code:403);
+        }
+
         abort_if(Gate::denies('task_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return view('tasks.edit', compact('task'));
@@ -58,6 +63,10 @@ class TasksController extends Controller
 
     public function destroy(Task $task)
     {
+        if ($task->user_id != auth()->id()){
+            abort(code:403);
+        }
+    
         abort_if(Gate::denies('task_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $task->delete();
